Load an image into njar

Load an image and optionally skeletonize it.

If you have already skeltonized the image externally, use load_skel().

Usage

load_image(p, skeletonize = TRUE, return_matrix = FALSE)

Arguments

p

The path to a binary map to load.

skeletonize

Whether to skeletonize the image using scikit-image through the reticulate package.

return_matrix

Whether to return the image as a matrix.

Value

An NJANet object or matrix containing the image data.

Binary image requirements

The provided image should be in .png, .jpg, or .bmp format.

Fully white pixels should indicate places where the network is present.

Examples

if (FALSE) { # interactive()
load_image("path/to/image.png")
}